# docs/swagger.yaml
basePath: /
definitions:
  auth.NewUser:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  auth.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  people.Person:
    properties:
      age:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "2.0"
paths:
  /:
    get:
      consumes:
      - '*/*'
      description: Get test on base path.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Show the status of server.
      tags:
      - Root Base
  /api/v1/people:
    get:
      description: Get people from db.
      parameters:
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all people.
      tags:
      - People
  /api/v1/person/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a person on DB.
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete a person.
      tags:
      - People
    get:
      consumes:
      - '*/*'
      description: Get person from db.
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get a single person.
      tags:
      - People
  /api/v1/person/{username}:
    put:
      consumes:
      - application/json
      description: Update a person data (age/email)
      parameters:
      - description: Person username
        in: path
        name: username
        required: true
        type: string
      - description: Person data to update
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/people.Person'
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update a person data
      tags:
      - People
  /api/v1/reservations:
    get:
      description: Get reservations from db.
      parameters:
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get all reservations.
      tags:
      - Reservations
  /api/v1/reservations/{id}:
    get:
      consumes:
      - '*/*'
      description: Get reservation from db.
      parameters:
      - default: Bearer
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get a single reservation.
      tags:
      - Reservations
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login form
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/auth.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login a person on DB.
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: New User to register
        in: body
        name: NewUser
        required: true
        schema:
          $ref: '#/definitions/auth.NewUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a person on DB.
      tags:
      - Authentication
swagger: "2.0"